♻ refactor: 추가 생성된 컴포넌트 index에 추가 작성
