+++ grass-diary/src/components/index.ts
@@ -12,3 +12,8 @@ export { default as Container } from './Container';
 export { EllipsisIcon, EllipsisBox } from './Ellipsis';
 export { default as Footer } from './NavBar/Footer';
 export { default as Comments } from './Comment/Comments';
+export { default as Menus } from './Button/Menus';
+export { default as Menu } from './Button/Menu';
+export { default as Modal } from './modal/Modal';
+export { default as CheckButton } from './modal/CheckButton';
+export { default as CustomButton } from './modal/CustomButton';
